ARROW-11200: [Rust] [DataFusion] Physical operators and expressions should have public accessor methods
Physical operators and expressions should have public accessor methods so that other crates can walk the physical plan and extract the values used to construct the operators and expressions.

This allows other projects to translate the physical plan into a different form or to serialize a physical plan to support distributed queries in a cluster.

Closes #9154 from andygrove/physical-plan-accessor-methods

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>