ARROW-11158: [Julia] Implement Decimal256 support for Julia
This PR also includes a few other bugfixes since the original Julia code
donation. The details of the release process of the code here and
transition from the JuliaData/Arrow.jl repository are still being ironed
out, so there has been development in both places and this synchronizes
the two.

cc: @nealrichardson

Closes #9121 from quinnj/jq/ARROW-11158

Authored-by: Jacob Quinn <quinn.jacobd@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>