ARROW-9634: [C++][Python] Restore non-UTC time zones when reading Parquet file that was previously Arrow
Currently we required an exact match for the time unit to restore the timezone, but I don't think there any problem in restoring the timezone also when the unit is different (the timezone is purely metadata, and should be transferable as is accross units).

Specifically, this solves the case of having nanosecond unit data as a start, which (by default) get stored as microsecond data in parquet, and thus currently didn't restore the timezone on roundtrip.

Closes #9336 from jorisvandenbossche/ARROW-9634-parquet-restore-timezone

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>