ARROW-9196: [C++][Compute] All casts accept scalar and sliced inputs
This is primarily a refactoring of the tests for "cast" to ensure that every case is also verified with scalar and sliced inputs through `CheckScalarUnary`. Caveat: ExtensionScalar is not implemented, so those checks aren't enabled. It should be straightforward for them to be enabled in a follow up, however.

NB: also resolves ARROW-9198

Closes #9490 from bkietz/9196-Make-temporal-casts-work-

Authored-by: Benjamin Kietzman <bengilgit@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>