ARROW-11580: [C++] Add CMake option ARROW_DEPENDENCY_SOURCE=VCPKG
This builds upon ARROW-11340 and makes it simpler to use vcpkg for dependencies.

Previously, to use vcpkg it was necessary to specify:
- `-DARROW_DEPENDENCY_SOURCE=SYSTEM`
- `-DCMAKE_TOOLCHAIN_FILE=/path/to/vcpkg.cmake`
- `-DVCPKG_TARGET_TRIPLET=xxx-yyy-zzz`
- Various`*_MSVC_*` variables on Windows
- Various `*_ROOT`, `*_LIB`, and `*_INCLUDE` variables in some cases
- `-D_VCPKG_INSTALLED_DIR=/path/to/vcpkg/installed/dir` in some cases

Now all that is necessary is:
- `-DARROW_DEPENDENCY_SOURCE=VCPKG`

Other variables are now optional and can be specified to override default behaviors.

After this is merged, I will make additional updates to the docs, focusing initially on Windows (ARROW-11336).

Closes #9553 from ianmcook/ARROW-11580-take2

Authored-by: Ian Cook <ianmcook@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>