ARROW-11560: [C++][FlightRPC] fix mutex error on SIGINT
Recently, interrupting a Python Flight server started aborting the process instead, because gRPC on Linux started using some code which is not signal-safe. This PR fixes that by spawning a separate thread that reads a byte from a self-pipe, then shuts down the server; the signal handler write to the pipe (which is a signal-safe operation).

This is a little janky, but fixes the issue. This still builds on Windows in my tests, but I couldn't actually import the resulting PyArrow to test it. However, PyArrow 3.0 via Conda still works on Windows. I'm not able to test on MacOS myself.

Closes #9633 from lidavidm/arrow-11560

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>