ARROW-12987: [C++][CI] Switch to bundled utf8proc with version 2.2 in Ubuntu 18.04 images
The tests are failing in the builds because the **utf8proc** dependency in Ubuntu 18.04 package is outdated.

I take a look at the utf8proc repository and I check that the error related to the build was fixed by that Pull Request https://github.com/JuliaStrings/utf8proc/pull/134 that is part of the 2.2 release: https://github.com/JuliaStrings/utf8proc/releases/tag/v2.2.0, but the Ubuntu 18.04 has only the 2.1.0 version in its package manager.

Closes #10468 from anthonylouisbsb/fixbug/fix-nightly-build-problems

Authored-by: Anthony Louis <anthony@simbioseventures.com>
Signed-off-by: Krisztián Szűcs <szucs.krisztian@gmail.com>